chore: Simplify Previously Sold on Artsy artworks rail
